+++ ai2_kit/core/cmd.py
@@ -1,0 +1,6 @@
+class CmdGroup:
+    """A group of commands"""
+
+    def __init__(self, items: dict, doc: str = '') -> None:
+        self.__doc__ = doc
+        self.__dict__.update(items)